+++ src/main/java/chess/domain/board/Board.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import chess.domain.piece.Bishop;
 import chess.domain.piece.EmptyPiece;
