+++ src/main/java/chess/domain/board/LocationDiff.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 public class LocationDiff {
