+++ src/main/java/chess/domain/board/Rank.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import java.util.Arrays;
 import java.util.Comparator;
