+++ src/main/java/chess/domain/piece/EmptyPiece.java
@@ -1,7 +1,7 @@
 package chess.domain.piece;
 
-import chess.domain.LocationDiff;
-import chess.domain.Direction;
+import chess.domain.board.LocationDiff;
+import chess.domain.board.Direction;
 
 public class EmptyPiece extends Piece{
     private static final double SCORE = 0;
