+++ src/test/java/chess/domain/LocationDiffTest.java
@@ -3,6 +3,8 @@ package chess.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.domain.board.Direction;
+import chess.domain.board.LocationDiff;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
