+++ src/test/java/chess/domain/state/DirectionTest.java
@@ -1,6 +1,6 @@
 package chess.domain.state;
 
-import chess.domain.Direction;
+import chess.domain.board.Direction;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
