+++ src/test/java/chess/domain/state/RunningTest.java
@@ -3,13 +3,10 @@ package chess.domain.state;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.Board;
-import chess.domain.File;
-import chess.domain.Location;
-import chess.domain.Rank;
-import chess.domain.piece.Pawn;
-import chess.domain.piece.Team;
-import java.util.Map;
+import chess.domain.board.Board;
+import chess.domain.board.File;
+import chess.domain.board.Location;
+import chess.domain.board.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
